Move CacheId idSite overwriting function to Context class.
